+++ app/widgets/Tabs/tabs.js
@@ -108,7 +108,6 @@ var Tabs = {
     }
 };
 
-
 document.addEventListener("DOMContentLoaded", function(event) {
     Tabs.create();
 });